+++ src/Database/Connectors/SqlServerConnector.php
@@ -2,6 +2,7 @@
 
 namespace DreamFactory\Core\SqlSrv\Database\Connectors;
 
+use DreamFactory\Core\Exceptions\InternalServerErrorException;
 use Illuminate\Database\Connectors\SqlServerConnector as LaravelSqlServerConnector;
 
 class SqlServerConnector extends LaravelSqlServerConnector
@@ -11,15 +12,16 @@ class SqlServerConnector extends LaravelSqlServerConnector
      */
     protected function getDsn(array $config)
     {
+        $drivers = $this->getAvailableDrivers();
         // We override the default usage of dblib, for the sqlsrv driver, now available on Linux as well.
-        if (in_array('sqlsrv', $this->getAvailableDrivers())) {
+        if (in_array('sqlsrv', $drivers)) {
             return $this->getSqlSrvDsn($config);
-        } elseif (in_array('dblib', $this->getAvailableDrivers())) {
+        } elseif (in_array('dblib', $drivers)) {
             return $this->getDblibDsn($config);
-        } elseif ($this->prefersOdbc($config)) {
+        } elseif (in_array('odbc', $drivers) && array_get_bool($config, 'odbc')) {
             return $this->getOdbcDsn($config);
         } else {
-            return $this->getDblibDsn($config);
+            throw new InternalServerErrorException('No acceptable driver for SQL Server found.');
         }
     }
 }