+++ src/Database/Schema/SqlServerSchema.php
@@ -5,22 +5,16 @@ use DreamFactory\Core\Database\Enums\DbFunctionUses;
 use DreamFactory\Core\Database\Enums\FunctionTypes;
 use DreamFactory\Core\Database\Schema\ColumnSchema;
 use DreamFactory\Core\Database\Schema\RoutineSchema;
-use DreamFactory\Core\Database\Components\Schema;
 use DreamFactory\Core\Database\Schema\TableSchema;
-use DreamFactory\Core\Enums\DbResourceTypes;
 use DreamFactory\Core\Enums\DbSimpleTypes;
 use DreamFactory\Core\Exceptions\ForbiddenException;
+use DreamFactory\Core\SqlDb\Database\Schema\SqlSchema;
 
 /**
  * Schema is the class for retrieving metadata information from a MS SQL Server database.
  */
-class SqlServerSchema extends Schema
+class SqlServerSchema extends SqlSchema
 {
-    /**
-     * Underlying database provides field-level schema, i.e. SQL (true) vs NoSQL (false)
-     */
-    const PROVIDES_FIELD_SCHEMA = true;
-
     const DEFAULT_SCHEMA = 'dbo';
 
     /**
@@ -29,6 +23,11 @@ class SqlServerSchema extends Schema
     const LEFT_QUOTE_CHARACTER = '[';
 
     const RIGHT_QUOTE_CHARACTER = ']';
+
+    public static function useSqlsrv()
+    {
+        return (in_array('sqlsrv', \PDO::getAvailableDrivers()));
+    }
 
     /**
      * @param boolean $refresh if we need to refresh schema cache.
@@ -43,16 +42,6 @@ class SqlServerSchema extends Schema
     /**
      * @inheritdoc
      */
-    public function getSupportedResourceTypes()
-    {
-        return [
-            DbResourceTypes::TYPE_TABLE,
-            DbResourceTypes::TYPE_VIEW,
-            DbResourceTypes::TYPE_PROCEDURE,
-            DbResourceTypes::TYPE_FUNCTION
-        ];
-    }
-
     protected function translateSimpleColumnTypes(array &$info)
     {
         // override this in each schema class
@@ -665,6 +654,35 @@ MYSQL;
     }
 
     /**
+     * @param $type
+     *
+     * @return mixed|null
+     */
+    public static function getNativeDateTimeFormat($type)
+    {
+        switch (strtolower(strval($type))) {
+            case DbSimpleTypes::TYPE_DATE:
+                return 'Y-m-d';
+
+            case DbSimpleTypes::TYPE_DATETIME:
+            case DbSimpleTypes::TYPE_DATETIME_TZ:
+                return 'Y-m-d H:i:s.u';
+
+            case DbSimpleTypes::TYPE_TIME:
+            case DbSimpleTypes::TYPE_TIME_TZ:
+                return 'H:i:s.u';
+
+            case DbSimpleTypes::TYPE_TIMESTAMP:
+            case DbSimpleTypes::TYPE_TIMESTAMP_TZ:
+            case DbSimpleTypes::TYPE_TIMESTAMP_ON_CREATE:
+            case DbSimpleTypes::TYPE_TIMESTAMP_ON_UPDATE:
+                return 'Y-m-d H:i:s.u P';
+        }
+
+        return null;
+    }
+
+    /**
      * Extracts the PHP type from DB type.
      *
      * @param ColumnSchema $column
@@ -740,7 +758,7 @@ MYSQL;
      */
     protected function getProcedureStatement(RoutineSchema $routine, array $param_schemas, array &$values)
     {
-        if (!in_array('sqlsrv', \PDO::getAvailableDrivers())) {
+        if (!self::useSqlsrv()) {
             // Note that using the dblib driver doesn't allow binding of output parameters,
             // and also requires declaration prior to and selecting after to retrieve them.
             $paramStr = '';
@@ -806,7 +824,7 @@ MYSQL;
 
     protected function doRoutineBinding($statement, array $paramSchemas, array &$values)
     {
-        if (!in_array('sqlsrv', \PDO::getAvailableDrivers())) {
+        if (!self::useSqlsrv()) {
             // do dblib version of binding
             foreach ($paramSchemas as $key => $paramSchema) {
                 switch ($paramSchema->paramType) {