+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.11.0] - 2017-07-27
+### Fixed
+- DF-1104 Fixed migrations and other issues with using pdo_dblib driver and SQL Server as system database
+- DF-1104 Overriding Laravel SQL Server handling to better support DBLIB driver
+- Separating base schema from SQL schema
+- Datetime settings handling
+
 ## [0.10.0] - 2017-06-05
 ### Changed
 - Cleanup - removal of php-utils dependency
@@ -60,7 +67,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.3.0 - 2016-05-27
 First official release working with the new [df-core](https://github.com/dreamfactorysoftware/df-core) library.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-sqlsrv/compare/0.10.0...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-sqlsrv/compare/0.11.0...HEAD
+[0.11.0]: https://github.com/dreamfactorysoftware/df-sqlsrv/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/dreamfactorysoftware/df-sqlsrv/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/dreamfactorysoftware/df-sqlsrv/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/dreamfactorysoftware/df-sqlsrv/compare/0.7.0...0.8.0